requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI/ML libraries
openai==1.3.0
google-generativeai==0.8.3

# HTTP and async
httpx==0.25.1
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Redis
redis==5.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
